Addition of a simClock node which controls time
In order to have finer control over the simulation time and the time-step, the `use_sim_time` parameter is now always set.
Depending on the value of `/world/timeControls` parameter, the simulation rate may be static (0), be triggerred by the `ctrlPWM` topic (1), or be free-running (2)

This change is in-line with the integration with ardupilot SITL function.