Fixes #34 on y-axis accelerations
When inert on the ground, the airplane would experience violent vibrations on the y-axis. These would be so high-frequency that they would not actually move the airplane but they would still get recorded on the accelerometer.

The cause was the ground reactions model. When modeling static friction, a division by the ground speed is performed. When ground speed is close to zero, this value grows very large. This results in very large friction values, which are capped, but due to a wrong coefficient, an oscillation would occur.

In the end, it was a filter coefficient instability problem.