Added support for APM SITL when running on `timeControls`=1
The value `timeControls`=1 forces the simulation to update ONLY when a new control input message is received by the UAV, right now published on the /<uav_name>/ctrlPWM topic.
In this way, the controller can drive the simulation at its own pace and NEVER miss a state measurement, due to lack of processing speed.

Essentially, the simulation clock is arbitrated by the controller.