Added a new friction model
This model only builds single-point contact points but NO wheels.
This model is force-based and not velocity based.
If a point is already moving, a constant force analogous to the normal force is applied.
If a point is not moving, then a friction force is applied to it, which equalizes the external forces, up to the stiction force value.