Abstracted the aerodynamics class
In suit with the changes in the propulsion model, now the aerodynamics model can have multiple instances on the same aircraft.
Moreover, they can now be tilted on one axis, like the motors.

This allows for modelling separate airfoils on an aircraft, eg the main wing, the horizontal and the vertical stabilizer, and placing them in their appropriate position on the airframe.

Configurations which benefit from this change are also the tilt-wing aircraft.